requirements.txt
pinecone-client
pinecone-text
pinecone-notebooks
langchain-community
python-dotenv
langchain_openai
